fix: :bug: family domain entities folder restructuring
updating test rules and business rules for family domain entities

errors when generating a family entity using the builder method AsHOF
